clean: move event to the src folder
